requirements.txt
social-auth-app-django==3.1.0
django-extensions==2.2.5
werkzeug==0.16.0
pyOpenSSL==19.0.0
easy-thumbnails==2.7
django-simple-captcha==0.5.14